refactor(renderer): move renderer test files to renderer test folder
